azure-ETL-EMR-AWS CDK-Power BI/Code/requirements.txt
aws-cdk-lib==2.51.1
constructs>=10.0.0,<11.0.0
nanoid==2.0.0
nanoid-dictionary==2.4.0
python-decouple==3.8